Move mergeParameters to new Util\ParametersManipulator class
